MAGETWO-30375: Magento/Customer/Controller/Adminhtml/Index/Save.php

+++ app/code/Magento/Customer/Model/Resource/CustomerRepository.php
@@ -168,32 +168,34 @@ class CustomerRepository implements \Magento\Customer\Api\CustomerRepositoryInte
         $this->customerRegistry->push($customerModel);
         $customerId = $customerModel->getId();
 
-        if ($customer->getId()) {
-            $existingAddresses = $this->getById($customer->getId())->getAddresses();
-            $getIdFunc = function($address) {
-                return $address->getId();
-            };
-            $existingAddressIds = array_map($getIdFunc, $existingAddresses);
-        } else {
-            $existingAddressIds = [];
-        }
-
-        $savedAddressIds = [];
-        foreach ($customer->getAddresses() as $address) {
-            $address = $this->addressBuilder
-                ->populate($address)
-                ->setCustomerId($customerId)
-                ->setRegion($address->getRegion())
-                ->create();
-            $this->addressRepository->save($address);
-            if ($address->getId()) {
-                $savedAddressIds[] = $address->getId();
+        if ($customer->getAddresses() !== null) {
+            if ($customer->getId()) {
+                $existingAddresses = $this->getById($customer->getId())->getAddresses();
+                $getIdFunc = function($address) {
+                    return $address->getId();
+                };
+                $existingAddressIds = array_map($getIdFunc, $existingAddresses);
+            } else {
+                $existingAddressIds = [];
             }
-        }
 
-        $addressIdsToDelete = array_diff($existingAddressIds, $savedAddressIds);
-        foreach ($addressIdsToDelete as $addressId) {
-            $this->addressRepository->deleteById($addressId);
+            $savedAddressIds = [];
+            foreach ($customer->getAddresses() as $address) {
+                $address = $this->addressBuilder
+                    ->populate($address)
+                    ->setCustomerId($customerId)
+                    ->setRegion($address->getRegion())
+                    ->create();
+                $this->addressRepository->save($address);
+                if ($address->getId()) {
+                    $savedAddressIds[] = $address->getId();
+                }
+            }
+
+            $addressIdsToDelete = array_diff($existingAddressIds, $savedAddressIds);
+            foreach ($addressIdsToDelete as $addressId) {
+                $this->addressRepository->deleteById($addressId);
+            }
         }
 
         $savedCustomer = $this->get($customer->getEmail(), $customer->getWebsiteId());
